Multi-level hierarchical category chooser
If there are hierarchical categories, the hierarchy breadcrumbs are
represented as the description text in the category select options.

This breadcrumb view is used by the modified version of jquery.select-
hierarchy to dynamically create multiple select widgets if javascript is
enabled.

If there's exactly one toplevel category and there are lower level
categories, only these are shown in the toplevel category chooser, and
the (single) toplevel select description is used as the toplevel select
prompt.