Log the index and active sort of clicked proposals.
When a user clicks a proposal the list index and active sort are logged
via request parameters in combination with the request log.  The option
'adhocracy.monitor_proposal_clicks' must be set to True to enable the
functionality. The special request parameters are stripped using a HTTP
redirect, to avoid having ugly URLs in the user's address bar.

This is needed for hhucn/adhocracy.hhu_theme#119.